modify for deployment

+++ controllers/factoryChildNodeController.js
@@ -1,5 +1,7 @@
 const FactoryChildNode = require('../models/factoryChildNode');
 const FactoryNode = require('../models/factoryNode');
+const Messages = require('../utils/ServerMessages');
+const NumberValidator = require('../utils/NumberValidator');
 
 module.exports = {
   getFactoryChildNodes(req, res) {
@@ -44,12 +46,18 @@ module.exports = {
       _id: nodeId,
     };
 
+    // validate params
     // set request data
     if (name) request.name = name;
-    if (minValue) request.minValue = minValue;
-    if (maxValue) request.maxValue = maxValue;
-    if (value) request.value = value;
-
+    if (minValue && NumberValidator.isPositiveInteger(minValue)) {
+      request.minValue = minValue;
+    }
+    if (maxValue && NumberValidator.isPositiveInteger(maxValue)) {
+      request.maxValue = maxValue;
+    }
+    if (value && NumberValidator.isPositiveInteger(value)) {
+      request.value = value;
+    }
     FactoryChildNode.findOneAndUpdate(filter, request, { new: true })
       .then((dbFactoryChildNode) => {
         req.app.io.emit('update node', dbFactoryChildNode);
@@ -69,10 +77,19 @@ module.exports = {
 
     const { factoryId } = req.params;
 
+    // validate params
+    if (!Array.isArray(nodeValues)) {
+      res.status(500).json(Messages.INVALID_PARAMETER_SPECIFIED);
+    }
+
     // set request data
     if (name) request.name = name;
-    if (minChildValue) request.minChildValue = minChildValue;
-    if (maxChildValue) request.maxChildValue = maxChildValue;
+    if (minChildValue && NumberValidator.isPositiveInteger(minChildValue)) {
+      request.minChildValue = minChildValue;
+    }
+    if (maxChildValue && NumberValidator.isPositiveInteger(maxChildValue)) {
+      request.maxChildValue = maxChildValue;
+    }
     if (!nodeValues) {
       res.status(500).json('No node values specified.');
     }
@@ -88,7 +105,7 @@ module.exports = {
     FactoryNode.findOne({ _id: factoryId })
       .then((dbFactory) => {
         // only add if adding these nodes to factory would result in <=15 child nodes
-        if (dbFactory.children.length + nodeValues.length < 15) {
+        if (dbFactory.children.length + nodeValues.length <= 15) {
           filteredNodes.forEach((nodeValue) => {
             // create a new node in the database for this node value
             const nodeObj = Object.assign({}, request, { value: nodeValue });
@@ -122,6 +139,7 @@ module.exports = {
     const {
       factoryId,
     } = req.params;
+
     const factoryFilter = {
       _id: factoryId,
     };
@@ -129,17 +147,21 @@ module.exports = {
     FactoryNode.findOneAndUpdate(factoryFilter, { children: [] })
       .then((dbFactoryNode) => {
         const nodeIds = dbFactoryNode.children;
-        nodeIds.forEach((nodeId) => {
-          const nodeFilter = {
-            _id: nodeId,
-          };
-          FactoryChildNode.findOneAndRemove(nodeFilter)
-            .then((dbFactoryChildNode) => {
-              req.app.io.emit('delete node', dbFactoryChildNode);
-              res.json(dbFactoryChildNode);
-            })
-            .catch(err => res.json(err));
-        });
+        if (nodeIds.length) {
+          nodeIds.forEach((nodeId) => {
+            const nodeFilter = {
+              _id: nodeId,
+            };
+            FactoryChildNode.findOneAndRemove(nodeFilter)
+              .then((dbFactoryChildNode) => {
+                req.app.io.emit('delete node', dbFactoryChildNode);
+                res.json(dbFactoryChildNode);
+              })
+              .catch(err => res.json(err));
+          });
+        } else {
+          res.end();
+        }
       })
       .catch(err => res.json(err));
   },
